requirements.txt
aioredis==1.3.1
appdirs==1.4.4
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.6
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
daphne==3.0.1
defusedxml==0.7.1
Django==3.1.7
django-channels==0.7.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-silk==4.1.0
django-templated-mail==1.1.1
django-unixtimestampfield==0.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
gprof2dot==2021.2.21
gunicorn==20.0.4
hiredis==1.1.0
hyperlink==21.0.0
idna==2.10
importlib-metadata==1.7.0
incremental==21.3.0
isodate==0.6.0
itypes==1.2.0
Jinja2==2.11.3
lxml==4.6.3
MarkupSafe==1.1.1
msgpack==1.0.2
oauthlib==3.1.0
Pillow==8.1.2
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.8.1
PyJWT==2.0.1
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-file==1.5.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
service-identity==18.1.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
toml==0.10.2
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
uritemplate==3.0.1
urllib3==1.26.4
zeep==4.0.0
zipp==3.4.1
zope.interface==5.3.0
